Added VisualTest back to tesselated_item
